Moved edition controllers to crew folder
